add hard-coded avatar to comments and temporary timestamp
